alter mod file

+++ cCompareFolder.psm1
@@ -1,10 +1,12 @@
 ﻿Function Compare-Folder
     {
     [CmdletBinding()]
-    Param($SourceFolder,$DestinationFolder)
+    Param($SourceFolder,$DestinationFolder,[switch]$Test)
 
     $source = Get-ChildItem $SourceFolder -Recurse
     $destination = Get-ChildItem $DestinationFolder
+
+
     
     try
         {
@@ -21,13 +23,16 @@
             $comp = $null
         }
 
+if ($Test)
+    {
     if ($null -ne $comp)
         {
-        switch($comp)
+        <#switch($comp)
             {
                 {$_.SideIndicator -eq "<="} {Write-Verbose "Source: $($_.InputObject) does not exist in $DestinationFolder";Copy-Item "$SourceFolder\$($_.InputObject)" -Destination $destinationFolder -Verbose}
                 {$_.SideIndicator -eq "=>"} {Write-Verbose "Destination: $($_.InputObject) does not exist in $SourceFolder";Remove-Item "$DestinationFolder\$($_.InputObject)" -Force -Verbose}
-            }
+            }#>
+            Write-Verbose "Differences Found"
         return $false
         }
     else
@@ -36,5 +41,25 @@
             return $true
         }
     }
+else
+    {
+    if ($null -ne $comp)
+        {
+        switch($comp)
+            {
+                {$_.SideIndicator -eq "<="} {Write-Verbose "Source: $($_.InputObject) does not exist in $DestinationFolder";Copy-Item "$SourceFolder\$($_.InputObject)" -Destination $destinationFolder -Verbose}
+                {$_.SideIndicator -eq "=>"} {Write-Verbose "Destination: $($_.InputObject) does not exist in $SourceFolder";Remove-Item "$DestinationFolder\$($_.InputObject)" -Force -Verbose}
+            }
+        #Write-Verbose "Differences Found"
+        #return $false
+        }
+    else
+        {
+            Write-Verbose "No differences detected between $SourceFolder and $DestinationFolder"
+        #return $true
+        }
+    }
+
+}
 
 Export-ModuleMember *
